pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "social-uploader"
version = "1.0.0"
description = "CLI tool to upload edited videos to YouTube, TikTok, and Instagram"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Social Uploader" }]
dependencies = [
    "click>=8.1.7,<9.0.0",
    "rich>=13.7.0,<14.0.0",
    "pyyaml>=6.0.1,<7.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "pydantic>=2.5.0,<3.0.0",
    "cryptography>=41.0.0,<43.0.0",
    "google-api-python-client>=2.108.0,<3.0.0",
    "google-auth-oauthlib>=1.1.0,<2.0.0",
    "google-auth-httplib2>=0.1.1,<0.2.0",
    "requests>=2.31.0,<3.0.0",
    "instagrapi>=2.0.0,<3.0.0",
    "opencv-python>=4.8.0,<5.0.0",
    "tenacity>=8.2.0,<9.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0,<8.0.0"]

[project.scripts]
social-uploader = "main:cli"

[tool.setuptools]
py-modules = ["main"]

[tool.setuptools.packages.find]
where = ["."]
include = ["uploaders*", "core*", "cli*"]

[tool.setuptools.package-data]
"*" = ["*.py"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
